requirements.txt
fastapi
uvicorn[standard]
google-generativeai
pydantic
python-dotenv